+++ static/components/f-code.js
@@ -1,10 +1,10 @@
 import {
   asset,
   getFieldName,
+  initField,
   oninvalid,
   pushOptions,
   updateField,
-  view,
 } from "./utils.js";
 import { Component } from "./component.js";
 import { init } from "../libs/code.js";
@@ -19,7 +19,7 @@ customElements.define(
       const name = getFieldName(this);
       const id = `field_${name}`;
 
-      view(this);
+      initField(this);
       dom("label", {
         for: id,
         html: schema.label,