+++ static/components/f-color.js
@@ -1,4 +1,4 @@
-import { getFieldName, oninvalid, updateField, view } from "./utils.js";
+import { getFieldName, initField, oninvalid, updateField } from "./utils.js";
 import { Component } from "./component.js";
 import dom from "dom";
 
@@ -11,7 +11,7 @@ customElements.define(
       const name = getFieldName(this);
       const id = `field_${name}`;
 
-      view(this);
+      initField(this);
       dom("label", {
         for: id,
         html: schema.label,