+++ static/components/f-datetime.js
@@ -1,9 +1,9 @@
 import {
   getFieldName,
+  initField,
   oninvalid,
   toLocal,
   updateField,
-  view,
 } from "./utils.js";
 import { Component } from "./component.js";
 import dom from "dom";
@@ -32,7 +32,7 @@ export class Input extends Component {
     const name = getFieldName(this);
     const id = `field_${name}`;
 
-    view(this);
+    initField(this);
     dom("label", {
       for: id,
       html: schema.label,