+++ static/components/f-radio.js
@@ -1,4 +1,4 @@
-import { getFieldName, labelify, oninvalid, view } from "./utils.js";
+import { getFieldName, initField, labelify, oninvalid } from "./utils.js";
 import { Component } from "./component.js";
 import dom from "dom";
 
@@ -11,7 +11,7 @@ customElements.define(
       const name = getFieldName(this);
       const id = `field_${name}`;
 
-      view(this);
+      initField(this);
       dom("span", { class: "field-label", html: schema.label }, this);
       const ul = dom("ul", { class: "field-check-list" }, this);