+++ static/components/f-relation.js
@@ -1,9 +1,9 @@
 import {
   getFieldName,
+  initField,
   oninvalid,
   pushOptions,
   updateField,
-  view,
 } from "./utils.js";
 import { Component } from "./component.js";
 import dom from "dom";
@@ -17,7 +17,7 @@ customElements.define(
       const name = getFieldName(this);
       const id = `field_${name}`;
 
-      view(this);
+      initField(this);
       dom("label", { for: id, html: schema.label }, this);
 
       dom(