+++ static/components/f-text.js
@@ -1,10 +1,10 @@
 import { Component } from "./component.js";
 import {
   getFieldName,
+  initField,
   oninvalid,
   pushOptions,
   updateField,
-  view,
 } from "./utils.js";
 import dom from "dom";
 
@@ -19,7 +19,7 @@ export class Input extends Component {
     const name = getFieldName(this);
     const id = `field_${name}`;
 
-    view(this);
+    initField(this);
     dom("label", { for: id, html: schema.label }, this);
 
     dom(