+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 - New option `previewURL` to return the URL of a file. It can be defined globally or per document and collection.
 - New option `sourcePath` to return the file of a URL.
+- New option `cssSelector` to highlight an element in the previewer.
 - `documentLabel` option to uploads.
 - New `edit` option to customize edit permissions for collections.
 - New `edit` option to customize edit permissions for documents.